Task02.

+++ src/se02/task02/Writeable.java
@@ -1,5 +1,5 @@
 package se02.task02;
 
 public interface Writeable {
-    double getCost();
+    int getCost();
 }
